Enable wildcard in the list of kernel functions
The handling of the -k option is too restrictive
as it stricly checks the names of the kernel functions
after it wrote them in /d/tracing/set_ftrace_filter.

However, a common usage of that function filtering with ftrace
is to use a wildcard character (*) that the ftrace tracer
automatically expands to all the matching kernel functions.

This enables the support of the wildcard (*) character
in the -k option to trace kernel functions.

Change-Id: Ifffae975c20e1c253157a3a6b44a14b4f342b9d0
Signed-off-by: Thomas Buhot <thomas.buhot@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>